requirements.txt
streamlit
autogen
openai
PyPDF2
python-docx
